# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Project Hub API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: student
    description: Student operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: User created, OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOtpRequest"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/personal-details:
    post:
      operationId: savePersonalDetails
      tags: [auth]
      summary: Save personal details after OTP verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonalDetailsRequest"
      responses:
        "200":
          description: Details saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout user
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /student/projects:
    get:
      operationId: getStudentProjects
      tags: [student]
      summary: Get all projects for the logged-in student
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [student]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /student/projects/{projectId}:
    get:
      operationId: getProject
      tags: [student]
      summary: Get a specific project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /student/projects/{projectId}/updates:
    get:
      operationId: getProjectUpdates
      tags: [student]
      summary: Get all updates for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectUpdate"
    post:
      operationId: addProjectUpdate
      tags: [student]
      summary: Add an update to a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddUpdateRequest"
      responses:
        "201":
          description: Update added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectUpdate"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /student/profile:
    get:
      operationId: getStudentProfile
      tags: [student]
      summary: Get student profile
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
    put:
      operationId: updateStudentProfile
      tags: [student]
      summary: Update student profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/dashboard:
    get:
      operationId: getAdminDashboard
      tags: [admin]
      summary: Get admin dashboard stats
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboardStats"

  /admin/projects:
    get:
      operationId: getAllProjects
      tags: [admin]
      summary: Get all student projects
      responses:
        "200":
          description: All projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectWithStudent"

  /admin/projects/{projectId}/status:
    put:
      operationId: updateProjectStatus
      tags: [admin]
      summary: Update project status
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users:
    get:
      operationId: getAllUsers
      tags: [admin]
      summary: Get all users
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"

  /admin/users/{userId}:
    delete:
      operationId: deleteUser
      tags: [admin]
      summary: Delete a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    SignupRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [student, admin]
      required:
        - email
        - username
        - password
        - role

    VerifyOtpRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        otp:
          type: string
      required:
        - email
        - otp

    PersonalDetailsRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        name:
          type: string
        mobile:
          type: string
        dob:
          type: string
        college:
          type: string
        department:
          type: string
          enum: [CSE, IT, AIDS, ECE, EEE, MECH]
        year:
          type: string
          enum: ["1st", "2nd", "3rd", "4th"]
      required:
        - email
        - name
        - mobile
        - dob
        - college
        - department
        - year

    LoginRequest:
      type: object
      properties:
        identifier:
          type: string
        password:
          type: string
      required:
        - identifier
        - password

    LoginResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: "#/components/schemas/UserResponse"
      required:
        - message
        - user

    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        username:
          type: string
        role:
          type: string
        name:
          type: string
          nullable: true
        mobile:
          type: string
          nullable: true
        dob:
          type: string
          nullable: true
        college:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        year:
          type: string
          nullable: true
        isVerified:
          type: boolean
      required:
        - id
        - email
        - username
        - role
        - isVerified

    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        dob:
          type: string
        college:
          type: string
        department:
          type: string
          enum: [CSE, IT, AIDS, ECE, EEE, MECH]
        year:
          type: string
          enum: ["1st", "2nd", "3rd", "4th"]

    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [Pending, Approved, Rejected, Completed]
        studentId:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - status
        - studentId
        - createdAt

    ProjectWithStudent:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [Pending, Approved, Rejected, Completed]
        studentId:
          type: integer
        createdAt:
          type: string
        studentName:
          type: string
          nullable: true
        studentUsername:
          type: string
      required:
        - id
        - title
        - description
        - status
        - studentId
        - createdAt
        - studentUsername

    CreateProjectRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
      required:
        - title
        - description

    ProjectUpdate:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        updateText:
          type: string
        date:
          type: string
      required:
        - id
        - projectId
        - updateText
        - date

    AddUpdateRequest:
      type: object
      properties:
        updateText:
          type: string
      required:
        - updateText

    UpdateProjectStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [Pending, Approved, Rejected, Completed]
      required:
        - status

    AdminDashboardStats:
      type: object
      properties:
        totalUsers:
          type: integer
        totalProjects:
          type: integer
        pendingProjects:
          type: integer
        completedProjects:
          type: integer
      required:
        - totalUsers
        - totalProjects
        - pendingProjects
        - completedProjects