# lib/api-zod/src/generated/types/adminDashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminDashboardStats {
  totalUsers: number;
  totalProjects: number;
  pendingProjects: number;
  completedProjects: number;
}