# lib/api-zod/src/generated/types/personalDetailsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { PersonalDetailsRequestDepartment } from "./personalDetailsRequestDepartment";
import type { PersonalDetailsRequestYear } from "./personalDetailsRequestYear";

export interface PersonalDetailsRequest {
  email: string;
  name: string;
  mobile: string;
  dob: string;
  college: string;
  department: PersonalDetailsRequestDepartment;
  year: PersonalDetailsRequestYear;
}