# lib/api-zod/src/generated/types/projectUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectUpdate {
  id: number;
  projectId: number;
  updateText: string;
  date: string;
}