# lib/api-zod/src/generated/types/projectWithStudent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectWithStudentStatus } from "./projectWithStudentStatus";

export interface ProjectWithStudent {
  id: number;
  title: string;
  description: string;
  status: ProjectWithStudentStatus;
  studentId: number;
  createdAt: string;
  studentName?: string | null;
  studentUsername: string;
}