# lib/api-zod/src/generated/types/updateProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateProfileRequestDepartment } from "./updateProfileRequestDepartment";
import type { UpdateProfileRequestYear } from "./updateProfileRequestYear";

export interface UpdateProfileRequest {
  name?: string;
  mobile?: string;
  dob?: string;
  college?: string;
  department?: UpdateProfileRequestDepartment;
  year?: UpdateProfileRequestYear;
}