# lib/api-zod/src/generated/types/userResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface UserResponse {
  id: number;
  email: string;
  username: string;
  role: string;
  name?: string | null;
  mobile?: string | null;
  dob?: string | null;
  college?: string | null;
  department?: string | null;
  year?: string | null;
  isVerified: boolean;
}